% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mun.R
\name{extract_mun}
\alias{extract_mun}
\title{Extract municipalities codes}
\usage{
extract_mun(identifier, string)
}
\arguments{
\item{identifier}{character. Alert identifier.}

\item{string}{character. String with municipalities codes.}
}
\value{
a tibble
}
\description{
Extract municipalities codes
}
