% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_alert.R
\name{parse_alert}
\alias{parse_alert}
\title{Fetch and parse alerts from INMET alerts RSS feed}
\usage{
parse_alert(alert_url)
}
\arguments{
\item{alert_url}{character. Alert url.}
}
\value{
a tibble with alert data
}
\description{
Fetch and parse alerts from INMET alerts RSS feed
}
