% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mun.R
\name{parse_mun}
\alias{parse_mun}
\title{Parse alerts by municipality codes}
\usage{
parse_mun(alerts, progress = TRUE)
}
\arguments{
\item{alerts}{tibble. Created with \code{parse_feed()}}

\item{progress}{logical. Show progress bar.}
}
\value{
a tibble with municipality codes and alerts data
}
\description{
Parse alerts by municipality codes
}
\examples{
alerts <- parse_feed()
parse_mun(alerts)
}
