% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_feed.R
\name{parse_feed}
\alias{parse_feed}
\title{Fetch and parse INMET alerts RSS feed}
\usage{
parse_feed(feed_url = NULL, progress = TRUE)
}
\arguments{
\item{feed_url}{character. INMET RSS feed. Defaults to current url.}

\item{progress}{logical. Show progress bar.}
}
\value{
a tibble with alerts data
}
\description{
Fetch and parse INMET alerts RSS feed
}
\examples{
parse_feed()
}
